requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
fastapi==0.104.1
idna==3.6
numpy==1.26.3
pandas==2.1.4
pydantic==2.4.2
pyswisseph==2.10.3.2
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
tzdata==2023.4
urllib3==2.1.0
uvicorn==0.23.2
a2wsgi==1.10.0
